backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==1.5.3
numpy==1.23.5
huggingface-hub==0.22.2
gunicorn==21.2.0